chapter3/artifacts/com.hbs.hub.WriteSenseHATSimulated/1.0.0/requirements.txt
awscrt==0.12.4
awsiotsdk==1.7.1
pkg-resources==0.0.0
